Add safety if dest dir does not exist

+++ gen_ipynb_from_py.py
@@ -71,6 +71,9 @@ def main():
         #print("Copying Notebooks to notebooks folder.")
         #print(glob.glob(f'*.ipynb'))
         for f in glob.glob(f'*.ipynb'):
+            if not os.path.exists(f'{notebooks_path}'):
+                print(f'Skipping {notebooks_path}/{f} as dest dir does not exist')
+                continue
             print(f'{scripts_path}/{f} -> {notebooks_path}/{f}')
             shutil.move(f'{scripts_path}/{f}', f'{notebooks_path}/{f}')
         if os.path.exists(f'{notebooks_path}/__init__.ipynb'):
